Security + big fixes

+++ backend/routes/user-router.js
@@ -2,29 +2,31 @@ const userRouter = require('express').Router()
 const User = require("../models/user-model")
 const List = require("../models/list-model")
 const Item = require("../models/item-model")
+const middleware = require("../middleware")
 
 //retrieve all existing lists ids and titles
-userRouter.get('/:username/lists', async (request, response) => {
+userRouter.get('/:username/lists', middleware.userExtractor, async (request, response) => {
     console.log("GET LISTS: ", request.params.username)
-    const user = await User.findOne({username: request.params.username})
-        .populate({
-            path: 'lists',
-            select: 'title'
-        })
-    response.status(200).json({lists: user.lists})
+    response.status(200).json({lists: request.user.lists})
 })
 
 //retrieve a list in detail
-userRouter.get('/:username/lists/:listId', async (request, response) => {
+userRouter.get('/:username/lists/:listId', middleware.userExtractor, async (request, response) => {
     console.log("GET LIST DETAILS: ", request.params.username)
     const list = await List.findById(request.params.listId).populate('items')
-    response.status(200).json(list)
+    if (request.user._id.equals(list.author)){
+        return response.status(200).json(list)
+    }
+    response.status(401).json({error: "Unauthorized access."})
 })
 
 //new list creation
-userRouter.post('/:username/lists', async (request, response) => {
+userRouter.post('/:username/lists', middleware.userExtractor, async (request, response) => {
     console.log("POST LISTS: ", request.params.username)
-    const user = await User.findOne({username: request.params.username})
+    const user = request.user
+    if (user.lists.length >= 10){
+        return response.status(400).json({error: "You have too many lists on this account."})
+    }
     const newList = new List({author: user._id})
     user.lists.push(newList)
     await newList.save()
@@ -33,10 +35,15 @@ userRouter.post('/:username/lists', async (request, response) => {
 })
 
 //list deletion
-userRouter.delete('/:username/lists/:listId', async (request, response) => {
+userRouter.delete('/:username/lists/:listId', middleware.userExtractor, async (request, response) => {
     console.log("DELETE LIST: ", request.params.username)
-    const user = await User.findOne({username: request.params.username})
+    const user = request.user
     const listId = request.params.listId
+    const list = await List.findById(listId)
+
+    list.items.forEach(async (curr) => {
+        await Item.findByIdAndDelete(curr.toString())
+    })
     await List.findByIdAndDelete(listId)
     user.lists = user.lists.filter(curr => {
         return curr._id.toString() !== listId
@@ -46,18 +53,29 @@ userRouter.delete('/:username/lists/:listId', async (request, response) => {
 })
 
 //list update
-userRouter.put('/:username/lists/:listId', async (request, response) => {
+userRouter.put('/:username/lists/:listId', middleware.userExtractor, async (request, response) => {
     console.log('UPDATE LIST TITLE AND DESC: ', request.params.username)
-    const newList = await List.findByIdAndUpdate(request.params.listId, request.body, {new: true}).populate('items')
-    response.status(200).json(newList)
+    const list = await List.findById(request.params.listId)
+    if (request.user._id.equals(list.author)){
+        const newList = await List.findByIdAndUpdate(request.params.listId, request.body, {new: true}).populate('items')
+        return response.status(200).json(newList)
+    }
+    response.status(401).json({error: "Unathorized access."})
 })
 
 //new item to list
-userRouter.post('/:username/lists/:listId/itemTransaction', async (request, response) => {
+userRouter.post('/:username/lists/:listId/itemTransaction', middleware.userExtractor, async (request, response) => {
     console.log('NEW ITEMS: ', request.params.username)
     const list = await List.findById(request.params.listId).populate('items')
+    if (!request.user._id.equals(list.author)){
+        return response.status(401).json({error: "Unathorized access."})
+    }
+    if (list.items.length >= 50){
+        return response.status(400).json({error: "You have too many items on this list."})
+    }
     const item = new Item({
         text: request.body.text,
+        author: request.user
     })
     await item.save()
     list.items.push(item)
@@ -66,9 +84,12 @@ userRouter.post('/:username/lists/:listId/itemTransaction', async (request, resp
 })
 
 //delete item from list
-userRouter.delete('/:username/lists/:listId/:itemId/itemTransaction', async (request, response) => {
+userRouter.delete('/:username/lists/:listId/:itemId/itemTransaction', middleware.userExtractor, async (request, response) => {
     console.log('DELETE ITEM: ', request.params.username)
     const list = await List.findById(request.params.listId).populate('items')
+    if (!request.user._id.equals(list.author)){
+        return response.status(401).json({error: "Unathorized access."})
+    }
     const id = request.params.itemId
     await Item.findByIdAndDelete(id)
     list.items = list.items.filter(curr => {
@@ -79,16 +100,16 @@ userRouter.delete('/:username/lists/:listId/:itemId/itemTransaction', async (req
 })
 
 //update item
-userRouter.put('/:username/lists/:listId/itemTransaction', async (request, response) => {
+userRouter.put('/:username/lists/:listId/itemTransaction', middleware.userExtractor, async (request, response) => {
     console.log('UPDATE ITEM: ',request.params.username)
-    console.log(request.body.done);
-    const list = await Item.findById(request.body.itemId)
-    list.done = request.body.done
-    console.log(list)
-    await list.save()
-    const newList = await Item.findById(request.body.itemId)
-    console.log(newList)
-    response.status(200).json({done: list.done})
+    const item = await Item.findById(request.body.itemId)
+    if (!request.user._id.equals(item.author)){
+        return response.status(401).json({error: "Unathorized access."})
+    }
+    item.done = request.body.done
+    await item.save()
+    await Item.findById(request.body.itemId)
+    response.status(200).json({done: item.done})
 })
 
 module.exports = userRouter
